Latest fixes

+++ GrooveOn.Models/ResponseObjects/IncomeByMonthResponse.cs
@@ -1,5 +1,5 @@
 public class IncomeByMonthResponse
 {
     public int Month { get; set; }
-    public float TotalIncome { get; set; }
+    public decimal TotalIncome { get; set; }
 }
